# docs/swagger.yaml
basePath: /api
definitions:
  models.Board:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateBoardRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.UpdateBoardRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Kanban application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kanban API Documentation
  version: "1.0"
paths:
...
  /organizations/{orgID}/projects/{projectID}/boards:
    get:
      description: Retrieves all boards within a specified project. User must own
        the project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of boards
          schema:
            items:
              $ref: '#/definitions/models.Board'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all boards in a project
      tags:
      - Boards
    post:
      consumes:
      - application/json
      description: Creates a new board within a specified project. User must own the
        project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board creation details
        in: body
        name: board
        required: true
        schema:
          $ref: '#/definitions/models.CreateBoardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Board created successfully
          schema:
            $ref: '#/definitions/models.Board'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new board
      tags:
      - Boards
  /organizations/{orgID}/projects/{projectID}/boards/{boardID}:
    delete:
      description: Deletes a specific board by its ID within a specified project.
        User must own the project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a board
      tags:
      - Boards
    get:
      description: Retrieves a specific board by its ID within a specified project.
        User must own the project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board details
          schema:
            $ref: '#/definitions/models.Board'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get board by ID
      tags:
      - Boards
    put:
      consumes:
      - application/json
      description: Updates a specific board by its ID within a specified project.
        User must own the project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      - description: Board update details
        in: body
        name: board
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBoardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Board updated successfully
          schema:
            $ref: '#/definitions/models.Board'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a board
      tags:
      - Boards
...
securityDefinitions:
  ApiKeyAuth:
    description: Enter your JWT token in the format "Bearer <your_token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...